backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.9
huggingface-hub==0.15.1
pandas==1.5.3
langchain==0.0.350
sentence-transformers==2.2.2
faiss-cpu==1.7.3
groq==0.2.0
python-dotenv==1.0.0